+++ lib/hetu_script.dart
@@ -14,7 +14,7 @@ export 'src/object.dart';
 export 'src/type.dart' show HTType;
 export 'src/namespace.dart' show HTNamespace;
 export 'src/class.dart' show HTClass;
-export 'src/vm/vm.dart';
+export 'src/bytecode/bytecode_interpreter.dart';
 export 'src/binding/external_class.dart';
 export 'src/binding/external_instance.dart';
 export 'src/binding/external_function.dart';